Mock buildIpfsConnection() in resume anchor requests tests
