+++ packages/cli/src/ceramic-daemon.ts
@@ -3,7 +3,7 @@ import express, { Request, Response } from 'express'
 import { Ceramic, CeramicConfig } from '@ceramicnetwork/core'
 import { RotatingFileStream } from '@ceramicnetwork/logger'
 import { ServiceMetrics as Metrics } from '@ceramicnetwork/observability'
-import { buildIpfsConnection } from './build-ipfs-connection.util.js'
+import { IpfsConnectionFactory } from './ipfs-connection-factory.js'
 import {
   DiagnosticsLogger,
   LoggerProvider,
@@ -271,7 +271,7 @@ export class CeramicDaemon {
   static async create(opts: DaemonConfig): Promise<CeramicDaemon> {
     const ceramicConfig = makeCeramicConfig(opts)
 
-    const ipfs = await buildIpfsConnection(
+    const ipfs = await IpfsConnectionFactory.buildIpfsConnection(
       opts.ipfs.mode,
       opts.network?.name,
       ceramicConfig.loggerProvider.getDiagnosticsLogger(),
